requirements.txt
duckdb==1.3.0
pandas==2.3.2
requests==2.32.5
python-dotenv==1.1.1
gtfs-realtime-bindings==1.0.0
protobuf==6.33.0
streamlit==1.55.0
plotly==6.3.1
